🚚 [RENAME] getAsset.ts 파일명 변경
